Changed csproj name from Domain.Services.Tests to Domain.Services.Impl.UnitTests, and moving files to 1.Domain folder
