--- 1.Domain/Domain.Services.Tests/Builders/IBuilder.cs
@@ -1,7 +1,0 @@
-﻿namespace Domain.Services.Tests.Builders
-{
-    public interface IBuilder<T> where T : class
-    {
-        T Build();
-    }
-}
