+++ 1.Domain/Domain.UnitTests/Builders/IBuilder.cs
@@ -1,0 +1,7 @@
+﻿namespace Domain.Services.Impl.UnitTests.Builders
+{
+    public interface IBuilder<T> where T : class
+    {
+        T Build();
+    }
+}
